CRAYSAT-1171: Implement simple CFS configuration utility
This commit adds the initial implementation of the CFS configuration
utility. At the moment, the options available for programmatic
configuration are just CFS configuration layers; this is used for
automating the process of adding or updating the SAT layer in CFS.

This commit adds a Python package, an executable, a Dockerfile and
scaffolding for an OCI container image, unit tests, and miscellaneous
documentation.

Test Description:
Fully build cfs-config-util image in Jenkins, and rebuild
sat-product-stream in Jenkins afterward. Test the installation process
on thanos and ensure that CFS layers are updated automatically at the
end of the install.sh flow.