CRAYSAT-1168: Add `version.sh` script to get version from changelog
We have been duplicating the version in `CHANGELOG.md` and `.version`,
which often leads to the two getting out of sync. The `.version` file is
just used by the dockerBuildPipeline to get the version for the created
docker image.

We already have a script that parses the changelog to get the latest
version it defines, and the `dockerBuildPipeline` supports specifying a
script to get the docker image version, so create `version.sh` script
that calls `changelog.py` against `CHANGELOG.md` to get the latest
version to specify as the docker image version.

Test Description:
Ran `version.sh` locally and verified that it got the right version
number. Examined the Jenkins pipeline build to ensure that the correct
docker image version was set.