CRAYSAT-1301: Find layer by clone URL and playbook instead of name
When searching for a layer to update or add in cfs-config-util, search
for layers by checking whether the path portion of the clone URL and the
playbook match rather than the name. The name is arbitrary, and admins
can set the name to whatever they want. The combination of the clone URL
and the playbook should more uniquely identify the layer that we intend
to add or replace.

Note that this considers a clone URL a match if just the path portion
matches. This was done because I've seen a system where the clone URL
was different than expected, perhaps due to some admin error. In that
case, the clone URL was of the form "https://user:pass@host/path". It
also seems possible that the hostname could change with coming DNS
changes. We can reconsider this behavior in the future.

Generalize the `ensure_layer` function to take a `state` parameter that
is specified as either 'present' or 'absent'. Then use this same
function in both `cfs_activate_version` and `cfs_deactivate_version`.

Implement new unit tests for ensuring a layer is present or absent in a
CFS configuration.

Test Description:
Unit tests pass. Pycodestyle passes.

Tested on lemondrop with a shell script that runs the `cfs-config-util`
image with podman in the same way as in `install.sh`. Tested updating a
layer with the name `sat-ncn` and a layer named `sat-2.2.17` to show
that the name no longer matters and instead matches on repo path and
playbook. Verified that the only changes made by `cfs-config-util` were
to the sat layer.