CRAYSAT-1220: Generalize cfs-config-util to support other products
Currently, `cfs-config-util` hard-codes values that are specific to the
SAT product, and it only supports one mode of operation, which is to
find the CFS configuration which applies to the master management NCNs,
and modify it in place by adding the layer for the SAT product.

Add options to the `cfs-config-util` entry point that generalize it and
allow it to be used by other products and in other workflows. These
options fit into three main categories:

* Options that describe a layer and whether that layer should be added
  or removed.
* Options that specify the base CFS configuration to modify. This may be
  a file, an existing CFS configuration, or HSM query parameters to find
  one or more existing CFS configurations.
* Options that specify how the modified CFS configuration should be
  saved, i.e., whether it should be saved in place, to a new
  configuration, or to a new file.

The intention is that other products will use this common functionality
in their product installers to present a consistent experience to the
admin performing the installation.

Also generalize the `cfs_[de]activate_version` functions in the
`cfs_config_util.activation` module. The `sat-install-utility` uses
these functions to activate or uninstall a version of the SAT product.
These functions are not fully generalized to the same extent as the CLI,
and more thought needs to be put into how to handle modification of CFS
configurations during product uninstall and activate.

Move the environment variables related to the API gateway into a
separate environment.py module that can be used in multiple places.

Add a dependency on the `cray-product-catalog` package which provides
the ability to query the product catalog ConfigMap.

These changes are not backwards compatible, so this requires a major
version bump.

Test Description:
New unit tests have been added. Existing unit tests have been modified.
All are passing. pycodestyle passes.

Tested by running with various options on baldar.