CRAYSAT-1418: Add LICENSE file and update every source file
Add a LICENSE file at the top level of the repository that includes the
MIT license text, and include that same license text in every source
file in the format expected by the license checker GitHub action.

Also update all copyright years to include the current year, 2022,
because this is required by the license checker.

Test Description:
Check if it passes the license checker.