CRAYSAT-1459: Make --base* args optional, fix help text
This change fixes the help text for the --base* options regarding their
behavior when no valid base is given. The text specifying that
cfs-config-util will start with an empty configuration has been deleted.
It is left implicit that non-existent configurations and files will
raise errors.

This change also modifies the behavior of cfs-config-util when no
--base* option is given. When no base is given, cfs-config-util will
start from an empty configuration with no layers. --save and
--save-suffix may not be used if no --base* is given. To avoid
clobbering existing configurations, --save-to-cfs and --save-to-file
will not work if an existing configuration or filename is given.

Test Description:

Run unit tests. On groot, manually do a sanity check by running
cfs-config-util commands inside container with various combinations of
options: without --base-* and with --save-* options, both once and
multiple times to check overwrite behavior; check with --base-* and with
--save-* to check for regressions; Check with non-existent --base-*
options.