CRAYSAT-1620: Use authentication when accessing csm-python-modules
This commit adds authenticated access when accessing the
csm-python-modules repository on artifactory.algol60.net. To do this,
a `.netrc` file containing read-only credentials is mounted into
the container at build time. This follows an example given by
the CASM devops team (See: CASMINST-5443).

This commit adds creation of a .netrc file on the Jenkins host on which the
build is running. The intention of this file is to provide credential
authentication in the case when the package's dependencies are being installed
in a VirtualEnv on the host to facilitate running unit tests. This uses the
utility jenkinsUtils.writeNetRc() found in csm-shared-library.

Test Description: Built image and package locally and in Jenkins