CRAYSAT-1750: Update PyYAML to 6.0.1
Update the version of PyYAML to 6.0.1 to resolve a build problem. This
problem was due to an incompatibility introduced in Cython 3.0, which
PyYAML addressed by pinning their builds to `Cython<3.0` for the time
being.

Update the versions of `cray-product-catalog` and `csm-api-client` to
newer versions which add PyYAML 6.0 compatibility.

Also remove `PyYAML`, `requests`, and `kubernetes` from the top-level
requirements of this repository. Those are no longer direct dependencies
of this repo since it uses the `csm-api-client` and
`cray-product-catalog` libraries for access to the CSM APIs and
Kubernetes now.

Since we don't use `PyYAML` directly, and the versions of our
dependencies have been updated for compatibility with PyYAML 6.0, we
don't have to worry about any incompatibilities with this code.

Test Description:
Test that build works in Jenkins.