CASMINST-6597: Add component update options to cfs-config-util
Add options to `cfs-config-util` to support modifying CFS components in
addition to modifying CFS configurations. This splits the main
`cfs-config-util` entry point into two different actions:
`update-configs` and `update-components`.

The `update-configs` action allows for the updating of a CFS
configuration, which includes adding and removing a layer or layers for
a product. This action is what the main entry point used to do, but with
some additional options added. These additional options add support for
backing up CFS configurations when they are overwritten, assigning a CFS
configuration to components after modification, making additional
changes to the impacted CFS components, and waiting on the CFS
components to become configured by the CFS Batcher.

The `update-components` action allows for updating CFS components and
waiting for them to reach a configured state. This is useful when you
need to update the configuration assigned to a component as its
`desiredConfig` or clear the error count or state of a component, but
you don't need to update the CFS configuration contents itself.

Note that because the main entry point now requires an action, this
change is backwards incompatible. If we want to update the other users
of this command (scripts in the SAT and COS product streams), we would
just need to update the call to include the `update-configs` action as
the first argument.

Update version of `csm-api-client` This branch needs new functionality
introduced in version 1.2.0 of the `csm-api-client` Python package.

Test Description:
Built against the new `csm-api-client` Python package and executed the
container with `podman` on mug with various options to modify CFS
configurations and update CFS components.

New unit tests pass.