CRAYSAT-1842: Add support for the CFS v3 API
Add support for the CFS v3 API. This pulls in a new version of the
csm-api-client library that adds support for CFS v3 while still
supporting CFS v2. The user of `cfs-config-util` can specify the CFS API
version using the new `--cfs-version` option.

This newer version of the library also fixes bugs with the handling of
CFS configurations that have specified `additional_inventory` and the
value `special_parameters.ims_require_dkms`. Older versions would drop
`additional_inventory` and `ims_require_dkms` when modifying it. This
will no longer occur with the newer version of `csm_api_client`.

Add a new the `--verbose` option to provide a way to enable debug
logging which will potentially be useful when debugging changes to this
script.

CRAYSAT-1842: Fix option handling