CRAYSAT-1919: Upgrade to nose2 for running unit tests
Upgrade to nose2 for running unit tests. The unit tests for this repo
run directly on the Jenkins runner, using the version of Python
installed there. That appears to have just been upgraded from Python 3.9
to 3.11 with a Debian 11 to 12 upgrade.

Ideally, we would be isolating our unit tests to run with a set version
of Python, e.g. the version of Python included in the container image
being built. For example, see how the sat repo runs unit tests in its
Dockerfile. However, upgrading to nose2 is good anyway, and this will be
a smaller, faster fix.

Test Description:
Ran unit tests with nose2 and verified that they passed. Will check that
the build works in Jenkins for the true test.

(cherry picked from commit c14a43ba5f074d9342043a291625ef856bb37b51)