CRAYSAT-1919: Fix 5.1.0 changelog entry
Move the updates of cray-product-catalog and csm-api-client dependencies
into 5.1.0 release of cfs-config-util. These are pretty significant
changes, so it makes sense to include them in the minor version bump
rather than a patch bump. This also simplifies backporting the other
versioning changes that need to be made to the release/5.0 branch.

I'm only changing the already published versions because the build
versions got messed up on artifactory anyway, and I'm going to be
creating a new 5.0.3 and 5.1.0 build to fix that.